apache_superset_extractor | :ok_hand: Updating code due to code review changes.
